# internal/server/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BuildMax API",
    "version": "0.0.7"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "conversation_id": { "type": "string" },
          "session_id": { "type": "string", "nullable": true },
          "status": { "type": "string" },
          "input": { "type": "string" },
          "title": { "type": "string" },
          "output": { "type": "string", "nullable": true },
          "created_by": { "type": "string" },
          "created_at": { "type": "integer" },
          "started_at": { "type": "integer", "nullable": true },
          "ended_at": { "type": "integer", "nullable": true },
          "error_message": { "type": "string", "nullable": true },
          "agent_id": { "type": "string", "nullable": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/teams/{team_id}/conversations/{conversation_id}/tasks": {
      "get": {
        "summary": "List tasks in a conversation",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "conversation_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create task in a conversation",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "conversation_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          }
        }
      }
    },
    "/api/teams/{team_id}/tasks/{task_id}": {
      "get": {
        "summary": "Get task",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "task_id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          }
        }
      }
    },
...
  }
}